pyproject.toml
[project]
name = "fynace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "cryptography>=46.0.3",
    "fastapi>=0.121.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "python-jose[cryptography]>=3.3.0",
    "plotly-express>=0.4.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
    "mercadopago>=2.3.0",
    "slowapi>=0.1.9",
    "supabase>=2.27.0",
]
